Big int -> SysInt move, and remove runtime_val

+++ minion/system/time_keeping.h
@@ -140,7 +140,7 @@ class TimerClass
   Output_Type output;
 public:
   
-  void setOutputType(int version)
+  void setOutputType(short version)
   {
     switch(version)
     {
